rename types.ts to errorClasses.ts
