+++ src/errorClasses.ts
