+++ src/rest/login.ts
@@ -1,7 +1,7 @@
 import { editUserInDb, getUserByEmailAndHash } from "../db/User/user";
 import { Request, Response } from "express";
 import { AccountNotFoundErrorJSON, createToken, ServerErrorJSON } from "../util";
-import { AuthenticationError } from "../types";
+import { AuthenticationError } from "../errorClasses";
 import StatusCode from "./status";
 import { User } from "../db/types";
 import { ObjectId } from "bson";
